detect/uint: Move uint logic into a separate module
This commit moves the uint logic into its own module.
